Production check: fix 43 confirmed findings across C++/Python/build/tests + zero-warning build
Iter 1 findings (13 MEDIUM):
- F-01: NaN bypasses epsilon guard in Cholesky (gauss_solver.cpp)
- F-02: Div-by-zero in Gaussian CDF when var=0 (gaussian_distribution.h)
- F-05: Kriging variance untested
- F-06: macOS arm64 Accelerate ILP64 not auto-detected (lapack_compat.h)
- F-09: Commented-out bounds check allows OOB (sugarbox_neighbour_lookup_distance.h)
- F-11: KI_SINGULARITY silently dropped from stats (cont_kriging.h)
- F-14: Cokriging mark1 bypasses range/angle validation (api.cpp)
- F-17: validate() accepts sill==0 (covariance_param.cpp)
- F-19: Unmapped byte values in INC remapping (api.cpp)
- F-21: MIN_RANGE=0.0 allows range==0 (validation.py) [partial]
- F-36/F-38: Missing golden-file and cross-path consistency tests
- F-37: OK/SK lack zero-sill guard (geo.py) [partial]

Iter 2 findings (30 MEDIUM):
- F-44: 64-bit seed truncated to 32-bit (bs_random_generator.h)
- F-49: Missing LVM means shape volume validation (api.cpp)
- F-50: Correlogram 2x cov() overhead (my_kriging_weights.h)
- F-53: Per-iteration probs vector allocation in SIS hot loop
- F-59/F-60/F-61: I/O error handling (load_property_from_file.cpp, property_writer.cpp)
- F-65/F-66: Python FFI crashes (validate_radius float, OSError in library search)
- F-68: NaN preserved in truncation loop (gtsim.py)
- F-69/F-70: O(n^2) numpy.append, triple-nested Python loop (routines.py)
- F-103: unknown_values set bypasses parser guard (geo.py)
- F-112/F-114: Build/CI fixes (cibuildwheel, cvariogram OpenMP)
- F-122-F-161: Test quality improvements (17 findings)
- F-159: Dead dependency pytest-benchmark removed

Warning cleanup: eliminated all 308 compiler warnings via override keywords,
static_cast for sign-compare, unused variable removal, and SYSTEM include for TNT.

Build: 0 errors, 0 warnings. Tests: 828 passed, 2 skipped (OpenMP).

+++ src/geo_bsd/_cvariogram/stack_layers.h
@@ -24,7 +24,7 @@ void stack_layers(
 
 	std::vector<double> cumulative_k(nx*ny, 0.0);
 
-	for(int layer = 0; layer < thick_layers.size(); layer++)
+	for(size_t layer = 0; layer < thick_layers.size(); layer++)
 	{
 		for(int i = 0; i < nx; i++)
 		{
